# lib/api-zod/src/generated/types/emailMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmailMessage {
  id: string;
  fromAddr: string;
  toAddr: string;
  headerSubject: string;
  receivedAt: string;
  text?: string;
  html?: string;
}